requirements.txt
a2a==0.44
instructor==1.9.0
openai==1.93.0
pydantic==2.11.7
pytest==8.3.5
PyYAML==6.0.2
PyYAML==6.0.2
Requests==2.32.4
tenacity==9.1.2
uvicorn==0.34.2
instructor==1.9.0
openai==1.93.0
Requests==2.32.4
schema==0.7.7
tenacity==9.1.2
utils==1.0.2
ichatbio-sdk==0.2.1